Added Citations Count in Expert Metrics
This commit adds citations count to the expert metrics. This looks at each expert's publications and then determines how many times all of their publications are cited by other experts within the context of a keywords search. It also provide the ability to sort the results by citations count (and changes citations count to be the default sorting metric).